rename 'copied' folder to 'firmware_headers'
