+++ Makefile
@@ -3,7 +3,7 @@ CROSS_COMPILE ?= arm-linux-gnueabi-
 PREFIX ?= $(CROSS_COMPILE)
 CC = $(PREFIX)gcc
 AR = $(PREFIX)ar
-SRCS = $(wildcard src/*.c) $(wildcard src/**/*.c) $(wildcard copied/**/*.c)
+SRCS = $(wildcard src/*.c) $(wildcard src/**/*.c) $(wildcard firmware_headers/**/*.c)
 OBJS = $(patsubst %.c,%.o,$(SRCS))
 CFLAGS += -fno-strict-aliasing -fwrapv
 CFLAGS += -Wall -Wextra -Wpointer-sign -Wno-unused-parameter
