+++ firmware_headers/bluetooth/hci_lib.h
