+++ firmware_headers/bluetooth/rfcomm.h
