+++ firmware_headers/bluetooth/sco.h
