+++ firmware_headers/lms2012/ev3_typedata.h
