+++ src/ev3_inputs/ev3_input_iic.c
@@ -7,7 +7,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 #include <string.h>
-#include "../../copied/lms2012/ev3_iic.h"
+#include "../../firmware_headers/lms2012/ev3_iic.h"
 #include "../ev3_wait.h"
 #include "../ev3_constants.h"
 #include "ev3_input_iic.h"
