+++ src/ev3_inputs/ev3_input_uart.c
@@ -8,7 +8,7 @@
 #include <stddef.h>
 #include <string.h>
 #include <unistd.h>
-#include "../../copied/lms2012/ev3_uart.h"
+#include "../../firmware_headers/lms2012/ev3_uart.h"
 #include "../ev3_wait.h"
 #include "../ev3_constants.h"
 #include "ev3_input_uart.h"
