+++ src/ev3_sensors/ev3_touch.c
@@ -1,4 +1,4 @@
-#include "../../copied/lms2012/ev3_basictypes.h"
+#include "../../firmware_headers/lms2012/ev3_basictypes.h"
 #include "../ev3_inputs/ev3_input_analog.h"
 #include "ev3_touch.h"
