+++ src/ev3_sensors/ht_compass.c
@@ -1,4 +1,3 @@
-#include "../../copied/lms2012/ev3_uart.h"
 #include "../ev3_inputs/ev3_input_iic.h"
 #include "../ev3_constants.h"
 #include "ht_compass.h"
